feat(refactor): split app into components
