+++ frontend/src/components/LazyImage/index.ts
@@ -1,0 +1,1 @@
+export { LazyImage } from './LazyImage';
