+++ frontend/src/components/PhotoFeed/index.ts
@@ -1,0 +1,2 @@
+export type { Photo } from "./PhotoFeed";
+export { PhotoFeed } from "./PhotoFeed";
