+++ frontend/src/components/UploadForm/index.ts
@@ -1,0 +1,1 @@
+export { UploadForm } from "./UploadForm";
